requirements.txt
# requirements.txt
pydicom
pyodbc
openpyxl
torch
torchvision
Pillow
